+++ include/PokemonDatas.h
@@ -1,0 +1,16 @@
+#pragma once
+
+
+using DataType = int;
+
+struct PokemonDatas
+{
+    DataType HP;                 // 宝可梦的生命值
+    DataType Attack;             // 宝可梦的攻击力
+    DataType Defense;            // 宝可梦的防御力
+    DataType SpecialAttack;      // 宝可梦的特殊攻击力
+    DataType SpecialDefense;     // 宝可梦的特殊防御力
+    DataType Speed;              // 宝可梦的速度
+};
+
+
