+++ test/test.vcxproj
@@ -150,6 +150,7 @@
     <ClInclude Include="..\include\Log.h" />
     <ClInclude Include="..\include\Pokemon.h" />
     <ClInclude Include="..\include\PokemonDataCalculation.h" />
+    <ClInclude Include="..\include\PokemonDatas.h" />
     <ClInclude Include="..\include\PokemonSkill.h" />
     <ClInclude Include="..\include\PokemonSpecies.h" />
     <ClInclude Include="..\include\PortalManager.h" />
@@ -179,6 +180,8 @@
     <ClCompile Include="..\src\main.cpp" />
     <ClCompile Include="..\src\Pokemon.cpp" />
     <ClCompile Include="..\src\PokemonDataCalculation.cpp" />
+    <ClCompile Include="..\src\PokemonDatas.cpp" />
+    <ClCompile Include="..\src\PokemonSpecies.cpp" />
     <ClCompile Include="..\src\PortalManager.cpp" />
     <ClCompile Include="..\src\ScreenDrawer.cpp" />
   </ItemGroup>
